requirements.txt
Django==2.1
djangorestframework==3.8.2
pytz==2018.5
PyYAML==3.13
